requirements.txt
discord.py>=2.3.2
openai>=1.12.0
requests>=2.31.0
gradio-client>=0.7.0
elevenlabs>=0.2.0
ccxt>=4.0.0
python-dotenv
aiosqlite
sqlalchemy[asyncio]
mcp>=1.0.0
Pillow>=10.0.0
httpx>=0.25.0
